+++ software/o_c_REV/HEM_ScaleDuet.ino
@@ -80,14 +80,14 @@ public:
         ResetCursor();
     }
         
-    uint32_t OnDataRequest() {
-        uint32_t data = 0;
+    uint64_t OnDataRequest() {
+        uint64_t data = 0;
         Pack(data, PackLocation {0,12}, mask[0]);
         Pack(data, PackLocation {12,12}, mask[1]);
         return data;
     }
 
-    void OnDataReceive(uint32_t data) {
+    void OnDataReceive(uint64_t data) {
         mask[0] = Unpack(data, PackLocation {0,12});
         mask[1] = Unpack(data, PackLocation {12,12});
 
@@ -185,10 +185,10 @@ void ScaleDuet_ToggleHelpScreen(bool hemisphere) {
     ScaleDuet_instance[hemisphere].HelpScreen();
 }
 
-uint32_t ScaleDuet_OnDataRequest(bool hemisphere) {
+uint64_t ScaleDuet_OnDataRequest(bool hemisphere) {
     return ScaleDuet_instance[hemisphere].OnDataRequest();
 }
 
-void ScaleDuet_OnDataReceive(bool hemisphere, uint32_t data) {
+void ScaleDuet_OnDataReceive(bool hemisphere, uint64_t data) {
     ScaleDuet_instance[hemisphere].OnDataReceive(data);
 }
