+++ software/o_c_REV/HEM_TrigSeq.ino
@@ -81,8 +81,8 @@ public:
         }
     }
 
-    uint32_t OnDataRequest() {
-        uint32_t data = 0;
+    uint64_t OnDataRequest() {
+        uint64_t data = 0;
         Pack(data, PackLocation {0,8}, pattern[0]);
         Pack(data, PackLocation {8,8}, pattern[1]);
         Pack(data, PackLocation {16,3}, end_step[0]);
@@ -90,7 +90,7 @@ public:
         return data;
     }
 
-    void OnDataReceive(uint32_t data) {
+    void OnDataReceive(uint64_t data) {
         pattern[0] = Unpack(data, PackLocation {0,8});
         pattern[1] = Unpack(data, PackLocation {8,8});
         end_step[0] = Unpack(data, PackLocation {16,3});
@@ -193,10 +193,10 @@ void TrigSeq_ToggleHelpScreen(bool hemisphere) {
     TrigSeq_instance[hemisphere].HelpScreen();
 }
 
-uint32_t TrigSeq_OnDataRequest(bool hemisphere) {
+uint64_t TrigSeq_OnDataRequest(bool hemisphere) {
     return TrigSeq_instance[hemisphere].OnDataRequest();
 }
 
-void TrigSeq_OnDataReceive(bool hemisphere, uint32_t data) {
+void TrigSeq_OnDataReceive(bool hemisphere, uint64_t data) {
     TrigSeq_instance[hemisphere].OnDataReceive(data);
 }
